refactor: move BlinkStickVariant to enums module
